feature(analytics): add phase-specific PR tracking and exercise history table
- **Personal Records**: Added `WorkoutPhase` (Combined, Concentric, Eccentric) to track peak force PRs separately for lifting and lowering phases (Issue #111).
- **Exercise Details**: Implemented a view mode toggle between Charts and a new tabular Session Data table. Added a Weight Trend chart.
- **Settings & Audio**: Added per-sound toggles for "Countdown Beeps" and "Rep Completion Sound" (Issue #100). Introduced distinct haptic/audio feedback for countdown ticks and warmup-to-working transitions.
- **Import/Export**: Introduced CSV import functionality for workout history with duplicate detection (Android implemented, iOS stubbed).
- **Volume & Estimation**: Added `BodyweightVolumeCalculator` for bodyweight exercise volume (Issue #229) and `RoutineTimeEstimator` using historical average set durations (Issue #225).
- **Preferences**: Added configurable weight increments (e.g., 0.1lb, 0.5kg) and an auto-start routine setting (Issues #266, #190).
- **Infrastructure**: Updated SQLDelight schema to version 20 to support PR phases and routine programming enhancements. Added various fake repositories for unit and E2E testing.

+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/domain/usecase/TrendAnalysisUseCase.kt
@@ -40,7 +40,7 @@ class TrendAnalysisUseCase {
         val ssTotal = yValues.sumOf { (it - yMean).pow(2).toDouble() }.toFloat()
         val ssResidual = xValues.zip(yValues).sumOf { (x, y) ->
             val predicted = slope * x + intercept
-            (y - predicted).pow(2).toDouble()
+            (y - predicted).pow(2)
         }.toFloat()
         val rSquared = if (ssTotal.compareTo(0f) > 0) {
             1f - (ssResidual / ssTotal)
